feat(Generators): ✨ Complete move of datetime and documentation
Move datetime data from data module to datetime module
Update documentation to include feature requirement.

Direct use of datetime generators will require the feature to be enabled
