+++ src/data/mod.rs
@@ -3,7 +3,6 @@ pub mod beer;
 pub mod color;
 pub mod computer;
 pub mod currency;
-pub mod datetime;
 pub mod files;
 pub mod language;
 pub mod log_level;
